Add clang-analyzer-security* checks
This adds clang-tidy's security checks to the codebase. The only one
requiring changes is the use of a floating point number as a loop
counter in src/gui/interface/Spinner.cpp. This is generally considered
bad practice, since floats often have mysterious off-by-one errors when
a decimal fraction is not representable as a binary fraction.