Use std::move where appropriate
std::move allows for the reuse of the contents of an object when
constructing a new one. A common example is when passing something by
value into a constructor of an object and using it to initialize a
member variable. Normally, this'll result in two copies, one to create
the parameter, and one to create the member variable from that param.
With std::move, this only becomes a copy into the parameter, while the
contents of the parameter are reused for the member variable.